pkgconfig file must not have non -I flags
The Cflags directive in .pc files is really meant for the
preprocessor, not for the compiler. It must only contain -D and -I
flags. Otherwise, it forces flags on downstream programs that may not
want them.

In case of the build.opensuse.org build farm, it has even led to
crashes as the machine that was randomly selected to build givaro had
SSE4, emitting -msse4.1 into Cflags, while the machine randomly
selected to build fflas-ffpack had only SSE2. That cannot work.

"Cflags: -I/usr/include  -fmessage-length=0 -grecord-gcc-switches -O2
-Wall -D_FORTIFY_SOURCE=2 -fstack-protector-strong -funwind-tables
-fasynchronous-unwind-tables  -mmmx -mpopcnt -msse -msse2 -msse3
-msse4.1 -msse4.2 -msse4a -mavx -mfma -mfma4 -mbmi -mfpmath=sse
-fabi-version=6"